requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
fastapi==0.124.4
fastembed==0.7.4
filelock==3.20.1
flatbuffers==25.9.23
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.3
MarkupSafe==3.0.3
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numpy==2.3.5
onnxruntime==1.23.2
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pillow==11.3.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
portalocker==3.2.0
protobuf==6.33.2
py_rust_stemmers==0.1.5
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
